requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipython==8.11.0
isort==5.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
